backend/requirements.txt
bidict             == 0.21.4
certifi            == 2021.10.8
cffi               == 1.14.5
charset-normalizer == 2.0.10
click              == 8.0.3
cloudpickle        == 1.6.0
colorama           == 0.4.4
Flask              == 2.0.2
Flask-SocketIO     == 5.1.1
gym                == 0.18.3
idna               == 3.3
itsdangerous       == 2.0.1
Jinja2             == 3.0.2
joblib             == 1.0.1
MarkupSafe         == 2.0.1
mkl-fft            == 1.3.0
mkl-random         == 1.1.0
mkl-service        == 2.3.0	
numpy              == 1.20.1
pickle-mixin       == 1.0.2
Pillow             == 8.2.0
pip                == 21.0.1
pycparser          == 2.20
pygame             == 2.0.1
pyglet             == 1.5.15
pyserial           == 3.5
python-engineio    == 4.3.1
python-socketio    == 5.5.1
requests           == 2.27.1
scikit-learn       == 0.24.1
scipy              == 1.6.1
setuptools         == 52.0.0.post20210125
six                == 1.15.0
threadpoolctl      == 2.1.0
torch              == 1.6.0
urllib3            == 1.26.8
Werkzeug           == 2.0.2
wheel              == 0.36.2
wincertstore       == 0.2